Fix cardzone having wrong view
Steps to reproduce:
1. Look at the top 2 cards of the library
2. Without closing, look at the top 3 cards of the library.
Result: when you move cards in and out of this zoneview, cards don't get
deleted or added to this zoneview.
Reason: in toggleZoneView first you create new zoneview, and only after
that old zoneview's destructor is called. In zoneview's destructor we
overwrite origzone's view pointer with NULL.
Solution: overwrite origzone's view pointer only if it is equal to the
zoneview being deleted.